refactor: 공통 필드(createdAt, updatedAt) 클래스 생성
