+++ src/main/java/nextstep/courses/domain/BaseEntity.java
@@ -1,0 +1,23 @@
+package nextstep.courses.domain;
+
+import java.time.LocalDateTime;
+
+public abstract class BaseEntity {
+
+    private LocalDateTime createdAt;
+    private LocalDateTime updatedAt;
+
+    public BaseEntity() {
+        this.createdAt = LocalDateTime.now();
+    }
+
+    public BaseEntity(LocalDateTime createdAt, LocalDateTime updatedAt) {
+        this.createdAt = createdAt;
+        this.updatedAt = updatedAt;
+    }
+
+    public LocalDateTime getCreatedAt() {
+        return createdAt;
+    }
+
+}
